auth/requirements.txt
flask==2.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
redis==4.0.2
PyJWT==2.4.0
python-dotenv==0.19.0
werkzeug==2.0.3